chore(List): convert List to TypeScript
