+++ packages/forma-36-react-components/src/components/List/List/List.tsx
@@ -1,24 +1,22 @@
-import React from 'react';
+import React, { Component } from 'react';
 import cn from 'classnames';
-import PropTypes from 'prop-types';
 import styles from './List.css';
 
-export class List extends React.Component {
-  static propTypes = {
-    extraClassNames: PropTypes.string,
-    children: PropTypes.node.isRequired,
-    testId: PropTypes.string,
-  };
+export interface ListProps {
+  extraClassNames?: string;
+  children: React.ReactNode;
+  testId?: string;
+}
 
+export class List extends Component<ListProps> {
   static defaultProps = {
-    extraClassNames: undefined,
     testId: 'cf-ui-list',
   };
 
   render() {
     const { extraClassNames, children, testId, ...otherProps } = this.props;
 
-    const classNames = cn(styles.List, extraClassNames);
+    const classNames = cn(styles['List'], extraClassNames);
 
     return (
       <ul className={classNames} data-test-id={testId} {...otherProps}>
